feat(core): add the append-only journal and its replay
Every accepted event reaches the journal before the store sees it, so a crash
costs at most the unforced window instead of everything the store had not yet
committed.

Three rules decide where the log ends, and each exists because of a way a
journal that looks correct is not. The checksum covers the whole header, not
just part of it. A frame records the position it was written at, so a frame read
from anywhere else is not a frame. And every frame carries its segment's salt,
so the leftovers of a recycled segment — which still have valid magic numbers
and valid checksums — cannot be replayed as live data. SQLite's own
write-ahead log uses the salt trick for the same reason.

The corruption property earned its keep immediately: flipping a bit at a random
offset found that the header's reserved bytes sat outside the checksummed
region, so a flip there went unnoticed. Reserved now sits before the checksum,
making the covered region one contiguous run, and the property flips bits at
every offset in the frame.

Frames also carry a low-water mark — the oldest capture timestamp not yet
journalled anywhere — which is what will bound a crash gap correctly. A gap
whose lower bound sat after events it must cover would leave a silent hole that
the refusal check could not see.

+++ trace-core/src/main/java/in/gravaxis/trace/core/geom/BlockBox.java
@@ -71,10 +71,10 @@ public record BlockBox(int minX, int minY, int minZ, int maxX, int maxY, int max
     }
 
     private static int clampHorizontal(long value) {
-        return (int) Math.clamp(value, EventRecords.MIN_HORIZONTAL, EventRecords.MAX_HORIZONTAL);
+        return Math.clamp(value, EventRecords.MIN_HORIZONTAL, EventRecords.MAX_HORIZONTAL);
     }
 
     private static int clampVertical(long value) {
-        return (int) Math.clamp(value, EventRecords.MIN_Y, EventRecords.MAX_Y);
+        return Math.clamp(value, EventRecords.MIN_Y, EventRecords.MAX_Y);
     }
 }
